fix(mcp-server): import routing constants in tools; use u32 suffix for NetworkId::with_suffix

+++ examples/kdapp-mcp-server/src/node_connector.rs
@@ -21,7 +21,7 @@ impl Default for NodeConfig {
         let network_id = match env::var("KDAPP_NETWORK").ok().as_deref() {
             Some("mainnet") => NetworkId::with_suffix(NetworkType::Mainnet, 0),
             Some(val) if val.starts_with("testnet-") => {
-                let suffix = val.trim_start_matches("testnet-").parse::<u8>().unwrap_or(10);
+                let suffix = val.trim_start_matches("testnet-").parse::<u32>().unwrap_or(10);
                 NetworkId::with_suffix(NetworkType::Testnet, suffix)
             }
             _ => NetworkId::with_suffix(NetworkType::Testnet, 10),
